+++ examples/p4est_3d_dgsem/elixir_advection_diffusion_nonperiodic.jl
@@ -58,6 +58,7 @@ boundary_conditions = BoundaryConditionDirichlet(initial_condition)
 semi = SemidiscretizationHyperbolicParabolic(mesh,
                                              (equations, equations_parabolic),
                                              initial_condition, solver;
+                                             solver_parabolic = ViscousFormulationBassiRebay1(),
                                              boundary_conditions = (boundary_conditions,
                                                                     boundary_conditions))