+++ NEWS.md
@@ -7,8 +7,25 @@ for human readability.
 
 ## Changes in the v0.15 lifecycle
 
+#### Added
+
+- It is now possible to use `ViscousFormulationLocalDG()` as the `solver_parabolic` for non-conforming `P4estMesh`es.
+This is useful for (locally) diffusion-dominated problems.
+This enables in particular adaptive mesh refinement for that solver-mesh combination ([#2712]).
 - Added functionality to `ScalarPlotData2D` allowing visualization a field provided by a user-defined scalar function ([#2796]).
 - Added `NonIdealCompressibleEuler2D` ([#2768]).
+- Generalization of `VolumeIntegralShockCapturingHG` and `VolumeIntegralShockCapturingRRG` to support different volume integrals on the
+  non-stabilized and stabilized elements/cells.
+  The generalized volume integral is called `VolumeIntegralShockCapturingHGType` and takes the three keyword arguments `volume_integral_default`,
+  `volume_integral_blend_high_order`, and `volume_integral_blend_low_order` besides the usual `indicator` argument.
+  In particular, `volume_integral_default` may be e.g.  `VolumeIntegralWeakForm` or `VolumeIntegralAdaptive`, i.e.,
+  the non-stabilized elements/cells are no longer restricted to `VolumeIntegralFluxDifferencing` only ([#2802]).
+- Added `IndicatorEntropyCorrection`. When combined with `VolumeIntegralAdaptive`, this blends together a stabilized and non-stabilized
+  volume integral based on the violation of a volume entropy condition. `IndicatorEntropyCorrectionShockCapturingCombined` additionally
+  guides the blending by taking the maximum of the entropy correction indicator and a shock capturing indicator ([#2764]).
+- The second-order subcell volume integral is no longer limited to reconstruction in primitive variables.
+  Instead, it is possible to reconstruct in custom variables, if functions `cons2recon` and `recon2cons` are provided to
+  `VolumeIntegralPureLGLFiniteVolumeO2` and `VolumeIntegralShockCapturingRRG`([#2817]).
 - Extended 3D support for subcell limiting with `P4estMesh` was added ([#2763]).
   In the new version, local (minimum and maximum) limiting for conservative variables (using the
   keyword `local_twosided_variables_cons` in `SubcellLimiterIDP()`) is supported.